Skipping tests properly if "gl2.integration.tests" property is not set.
