+++ integration-tests/src/test/java/integration/BaseRestTest.java
@@ -20,6 +20,7 @@ import org.junit.ClassRule;
 import org.junit.Rule;
 
 public class BaseRestTest extends BaseRestTestHelper {
+    @ClassRule public static RunIfProperty runIfProperty = new RunIfProperty("gl2.integration.tests");
     @ClassRule public static RestAssuredSetupRule restAssuredSetupRule = new RestAssuredSetupRule();
     @Rule public RequiredVersionRule requiredVersionRule = new RequiredVersionRule(restAssuredSetupRule);
     @Rule public MongoDbSeedRule mongoDbSeedRule = new MongoDbSeedRule();
